Problem: Prod builds of depending apps are broken
Solution:

The error from depending apps we were getting was:

```
== Compilation error in file lib/uptight/base.ex ==
** (ArgumentError) could not load module Jason.Encoder due to reason :unavailable
    (elixir 1.13.4) lib/protocol.ex:315: Protocol.assert_protocol!/2
    lib/uptight/base.ex:136: (file)
    (elixir 1.13.4) lib/kernel/parallel_compiler.ex:346: anonymous fn/5 in Kernel.ParallelCompiler.spawn_workers/7
could not compile dependency :uptight, "mix compile" failed. Errors may have been logged above. You can recompile this dependency with "mix deps.compile uptight", update it with "mix deps.update uptight" or clean it with "mix deps.clean uptight"
```

We assume at the moment that it's because we don't list Jason as a dependency, and since apps have Jason as a transient, dev and test somehow worked, whereas prod didn't.

That's why we add jason as a dep to uptight, publish, and hope for best with the new version.